requirements.txt
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
Flask==3.0.2
Flask-SocketIO==5.3.6
python-engineio==4.9.0
python-socketio==5.11.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
email-validator==2.1.0.post1